add rvalue_arguments_tests.cpp as suggested by Alex Merry to test the fix of issue #23
